+++ src/components/member/tabs/quote-tab/common.ts
@@ -1,4 +1,5 @@
 import styled from 'react-emotion'
+import { colorsV2 } from '@hedviginsurance/brand/index'
 
 export const ErrorMessage = styled('pre')({
   paddingTop: '1rem',
@@ -9,4 +10,10 @@ export const BottomSpacerWrapper = styled('div')({
 })
 export const Muted = styled('div')({
   opacity: 0.8,
+})
+export const ActionsWrapper = styled('div')({
+  background: colorsV2.flamingo200,
+  padding: '1rem',
+  width: '100%',
+  marginBottom: '1rem',
 })
